RANGER-1403:There is a problem in buildks class when delete invalid keystore file.

+++ credentialbuilder/src/main/java/org/apache/ranger/credentialapi/buildks.java
@@ -535,7 +535,7 @@ public class buildks {
 			}
 			if(keystore!=null && !keystore.isEmpty()){
 				File file =new File(keystore);
-				if(file!=null && file.length()==0){
+				if(file!=null && file.exists() && file.length()==0){
 					System.out.println("Provider file '"+keystore+"' is in invalid state or corrupt!! will try to delete first.");
 					file.delete();
 					file=null;
